Create change bgcolor component
